move to map body
